Move operations.py to a more unique name
